requirements.txt
beautifulsoup4
fastapi
httpx
ipython
numpy
pandas
prometheus-client
python-dotenv
requests
scikit-learn
structlog
uvicorn
